system_code/requirements.txt
Django>=4.2
djangorestframework>=3.14
torch>=2.0
transformers>=4.30
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
jieba>=0.42
sentence-transformers>=2.2
